+++ randovania/layout/trick_level.py
@@ -160,4 +160,5 @@ class TrickLevelConfiguration(BitPackValue):
             15,  # Instant Morph
             26,  # Invisible Objects
             27,  # Standable Terrain
+            28,  # Terminal Fall Abuse
         })