+++ test/layout/test_permalink.py
@@ -25,7 +25,7 @@ def test_encode(mock_dictionary_byte_hash: MagicMock):
 
     # Assert
     mock_dictionary_byte_hash.assert_called_once_with(link.layout_configuration.game_data)
-    assert encoded == "gAAAfRePAAABYYAAAHNAvA=="
+    assert encoded == "gAAAfRePAAAAsMAAADmgmA=="
 
 
 @pytest.mark.parametrize("invalid", [
@@ -94,7 +94,7 @@ def test_decode(mock_dictionary_byte_hash: MagicMock):
 
     # This test should break whenever we change how permalinks are created
     # When this happens, we must bump the permalink version and change the tests
-    encoded = "gAAAfReLCAABcYAAAHNAtw=="
+    encoded = "gAAAfReLCAAAuMAAADmgpw=="
 
     expected = Permalink(
         seed_number=1000,
@@ -141,7 +141,7 @@ def test_decode_mock_other(mock_packer_unpack: MagicMock,
     mock_layout_unpack.return_value.game_data = {"test": True}
 
     # Uncomment this line to quickly get the new encoded permalink
-    # assert expected.as_str == ""
+    #assert expected.as_str == ""
     # print(expected.as_str)
 
     # Run